quantity = models.PositiveIntegerField

+++ app/models.py
@@ -4,7 +4,7 @@ class App(models.Model):
 
     nomenclature = models.CharField('Номенклатура', max_length=255)
     unit = models.CharField('Единица', max_length=10)
-    quantity = models.DecimalField('Количество', max_digits=10, decimal_places=2, null=True, blank=True)
+    quantity = models.PositiveIntegerField('Количество', null=True, blank=True)
 
     class Meta:
         verbose_name = 'Склад инструмента'
